requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
ebooklib==0.18
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
deep-translator>=1.11
bleach==6.1.0
celery==5.3.4
redis==5.0.1
django-celery-results==2.6.0
django-celery-beat==2.5.0
python-decouple==3.8
drf-yasg==1.21.7
django-filter==23.5
ao3-api>=2.3.0
